Adjust encodings, tests
Writing to a bytes.Buffer now gets a minor shortcut in that most types,
except for arrays, will skip hitting a temporary buffer. Arrays are the
special case, as any element with an error must forbid writing any part
of the array (since I had an unwritten goal of keeping unbuffered
writes atomic).

In addition, niltype has been renamed to nilmsg and all tests are run
once against a bytes.Buffer to trigger the shortcut and once against
ioutil.Discard to trigger the temp buffer path.

Change-Id: I3649c9285ae40b42939734f3295b620fea699a4d